[EI-264] [api] fix: 테스트 타임아웃 시간 수정

+++ apps/api/src/numerical-guidance/infrastructure/adapter/persistence/indicator-board-metadata/indicator-board-metadata.persistent.adapter.ts
@@ -342,8 +342,19 @@ export class IndicatorBoardMetadataPersistentAdapter
   async deleteIndicatorBoardMetadata(id: string) {
     try {
       const indicatorBoardMetadataEntity: IndicatorBoardMetadataEntity =
-        await this.indicatorBoardMetadataRepository.findOneBy({ id: id });
+        await this.indicatorBoardMetadataRepository.findOne({
+          relations: ['member'],
+          where: {
+            id,
+          },
+        });
       this.nullCheckForEntity(indicatorBoardMetadataEntity);
+
+      const { member } = indicatorBoardMetadataEntity;
+      const metadatas = await this.loadIndicatorBoardMetadataList(member.id);
+      if (metadatas.length <= 1) {
+        throw new BadRequestException();
+      }
 
       await this.indicatorBoardMetadataRepository.remove(indicatorBoardMetadataEntity);
     } catch (error) {
@@ -361,6 +372,13 @@ export class IndicatorBoardMetadataPersistentAdapter
           1. id 값이 uuid 형식을 잘 따르고 있는지 확인해주세요.`,
           message: '정보를 불러오는 중에 문제가 발생했습니다. 다시 시도해주세요.',
           cause: error,
+        });
+      } else if (error instanceof BadRequestException) {
+        throw new BadRequestException({
+          HttpStatus: HttpStatus.BAD_REQUEST,
+          error: `[ERROR] 지표보드 메타데이터를 삭제하는 도중에 entity 오류가 발생했습니다.
+          2. 메타데이터가 1개 미만으로 감소하는지 확인해주세요.`,
+          message: '정보를 불러오는 중에 문제가 발생했습니다. 다시 시도해주세요.',
         });
       } else {
         throw new InternalServerErrorException({
